refactor: move pixi components to seperate folders
