+++ components.d.ts
@@ -1,19 +1,19 @@
 // generated by unplugin-vue-components
 // We suggest you to commit this file into source control
 // Read more: https://github.com/vuejs/core/pull/3399
-import '@vue/runtime-core';
+import '@vue/runtime-core'
 
-export {};
+export {}
 
 declare module '@vue/runtime-core' {
   export interface GlobalComponents {
-    AnimatorPreview: typeof import('./src/components/worldeditor/AnimatorPreview.vue')['default'];
-    PixiApplication: typeof import('./src/components/pixi/PixiApplication.vue')['default'];
-    PixiAssets: typeof import('./src/components/pixi/PixiAssets.vue')['default'];
-    PixiLoader: typeof import('./src/components/pixi/PixiLoader.vue')['default'];
-    PixiTicker: typeof import('./src/components/pixi/PixiTicker.vue')['default'];
-    PixiViewport: typeof import('./src/components/pixi/viewport/PixiViewport.vue')['default'];
-    RouterLink: typeof import('vue-router')['RouterLink'];
-    RouterView: typeof import('vue-router')['RouterView'];
+    AnimatorPreview: typeof import('./src/components/worldeditor/AnimatorPreview.vue')['default']
+    PixiApplication: typeof import('./src/components/pixi/app/PixiApplication.vue')['default']
+    PixiAssets: typeof import('./src/components/pixi/assets/PixiAssets.vue')['default']
+    PixiLoader: typeof import('./src/components/pixi/loader/PixiLoader.vue')['default']
+    PixiTicker: typeof import('./src/components/pixi/ticker/PixiTicker.vue')['default']
+    PixiViewport: typeof import('./src/components/pixi/viewport/PixiViewport.vue')['default']
+    RouterLink: typeof import('vue-router')['RouterLink']
+    RouterView: typeof import('vue-router')['RouterView']
   }
 }
