+++ src/components/pixi/app/types.ts
@@ -1,0 +1,5 @@
+import * as PIXI from 'pixi.js';
+
+export type PixiApplicationInstance = {
+  application: PIXI.Application;
+};
