+++ src/components/pixi/assets/PixiAssets.spec.ts
