+++ src/components/pixi/assets/PixiAssets.vue
