+++ src/components/pixi/assets/types.ts
@@ -1,0 +1,5 @@
+import { AssetsClass } from '@pixi/assets';
+
+export type PixiAssetsInstance = {
+  assets: AssetsClass;
+};
