+++ src/components/pixi/loader/PixiLoader.vue
