+++ src/components/pixi/loader/types.ts
@@ -1,0 +1,5 @@
+import * as PIXI from 'pixi.js';
+
+export type PixiLoaderInstance = {
+  loader: PIXI.Loader;
+};
