+++ src/components/pixi/ticker/PixiTicker.vue
