+++ src/components/pixi/ticker/types.ts
@@ -1,0 +1,5 @@
+import * as PIXI from 'pixi.js';
+
+export type PixiTickerInstance = {
+  ticker: PIXI.Ticker;
+};
